move workflow code to separate module
